# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pickleball Club Ladder API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /seasons:
    get:
      operationId: listSeasons
      tags: [seasons]
      summary: List all seasons
      responses:
        "200":
          description: List of seasons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Season"
    post:
      operationId: createSeason
      tags: [seasons]
      summary: Create a new season (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSeasonBody"
      responses:
        "201":
          description: Season created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Season"
  /seasons/active:
    get:
      operationId: getActiveSeason
      tags: [seasons]
      summary: Get active season
      responses:
        "200":
          description: Active season
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Season"
        "404":
          description: No active season
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /seasons/{id}:
    get:
      operationId: getSeason
      tags: [seasons]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Season
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Season"
  /seasons/{id}/activate:
    post:
      operationId: activateSeason
      tags: [seasons]
      summary: Activate a season (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Season activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Season"
  /seasons/{id}/deactivate:
    post:
      operationId: deactivateSeason
      tags: [seasons]
      summary: Deactivate a season (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Season deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Season"

  # TEAMS
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List teams
      parameters:
        - name: season_id
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamWithPlayers"
  /teams/my-team:
    get:
      operationId: getMyTeam
      tags: [teams]
      summary: Get the current user's team for the active season
      responses:
        "200":
          description: My team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamWithPlayers"
        "404":
          description: No team found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamWithPlayers"
  /teams/{id}/matches:
    get:
      operationId: getTeamMatches
      tags: [teams]
      summary: Get match history for a team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Team matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchWithDetails"
  /teams/{id}/status:
    patch:
      operationId: updateTeamStatus
      tags: [teams]
      summary: Update team status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamStatusBody"
      responses:
        "200":
          description: Team updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamWithPlayers"

  # INVITATIONS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Player:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [player, admin]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - fullName
        - email
        - role
        - isActive
        - createdAt
...
    Season:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - startDate
        - endDate
        - isActive
        - createdAt
    CreateSeasonBody:
      type: object
      properties:
        name:
          type: string
        startDate:
          type: string
        endDate:
          type: string
      required:
        - name
        - startDate
        - endDate

    # Team
    Team:
      type: object
      properties:
        id:
          type: string
        seasonId:
          type: string
        player1Id:
          type: string
        player2Id:
          type: string
        teamName:
          type: string
        status:
          type: string
          enum: [pending, active, inactive]
        createdAt:
          type: string
      required:
        - id
        - seasonId
        - player1Id
        - player2Id
        - teamName
        - status
        - createdAt
    TeamWithPlayers:
      allOf:
        - $ref: "#/components/schemas/Team"
        - type: object
          properties:
            player1:
              $ref: "#/components/schemas/Player"
            player2:
              $ref: "#/components/schemas/Player"
            standing:
              $ref: "#/components/schemas/LadderStanding"
            season:
              $ref: "#/components/schemas/Season"
    UpdateTeamStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, active, inactive]
      required:
        - status

    # Invitation
...
    LadderStanding:
      type: object
      properties:
        id:
          type: string
        seasonId:
          type: string
        teamId:
          type: string
        position:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        lastMatchDate:
          type: string
        createdAt:
          type: string
      required:
        - id
        - seasonId
        - teamId
        - position
        - wins
        - losses
        - createdAt
...
    Challenge:
      type: object
      properties:
        id:
          type: string
        seasonId:
          type: string
        challengerTeamId:
          type: string
        challengedTeamId:
          type: string
        status:
          type: string
          enum: [pending, accepted, scheduling, scheduled, completed, cancelled, disputed]
        createdAt:
          type: string
      required:
        - id
        - seasonId
        - challengerTeamId
        - challengedTeamId
        - status
        - createdAt
    ChallengeWithTeams:
      allOf:
        - $ref: "#/components/schemas/Challenge"
        - type: object
          properties:
            challengerTeam:
              $ref: "#/components/schemas/TeamWithPlayers"
            challengedTeam:
              $ref: "#/components/schemas/TeamWithPlayers"
            match:
              $ref: "#/components/schemas/MatchWithDetails"
            myTeamId:
              type: string
            overlappingSlots:
              type: array
              items:
                $ref: "#/components/schemas/TimeSlot"
            challengerAvailabilitySubmitted:
              type: boolean
            challengedAvailabilitySubmitted:
              type: boolean
...
    TimeSlot:
      type: object
      properties:
        date:
          type: string
        times:
          type: array
          items:
            type: string
      required:
        - date
        - times

    # Availability
...
    Match:
      type: object
      properties:
        id:
          type: string
        challengeId:
          type: string
        scheduledDate:
          type: string
        scheduledTime:
          type: string
        courtLocation:
          type: string
        bookedByTeamId:
          type: string
        status:
          type: string
          enum: [scheduled, completed, disputed]
        createdAt:
          type: string
      required:
        - id
        - challengeId
        - scheduledDate
        - scheduledTime
        - courtLocation
        - status
        - createdAt
    MatchScore:
      type: object
      properties:
        id:
          type: string
        matchId:
          type: string
        gameNumber:
          type: integer
        team1Score:
          type: integer
        team2Score:
          type: integer
      required:
        - id
        - matchId
        - gameNumber
        - team1Score
        - team2Score
    MatchResult:
      type: object
      properties:
        id:
          type: string
        matchId:
          type: string
        winnerTeamId:
          type: string
        loserTeamId:
          type: string
        submittedByTeamId:
          type: string
        confirmedByTeamId:
          type: string
        submittedAt:
          type: string
        confirmedAt:
          type: string
        autoConfirmed:
          type: boolean
        disputeReason:
          type: string
        disputeResolved:
          type: boolean
      required:
        - id
        - matchId
        - winnerTeamId
        - loserTeamId
        - submittedByTeamId
        - submittedAt
        - autoConfirmed
        - disputeResolved
    MatchWithDetails:
      allOf:
        - $ref: "#/components/schemas/Match"
        - type: object
          properties:
            challenge:
              $ref: "#/components/schemas/ChallengeWithTeams"
            scores:
              type: array
              items:
                $ref: "#/components/schemas/MatchScore"
            result:
              $ref: "#/components/schemas/MatchResult"
...